[update] エディタウィンドウ、通信用クライアント、チケットデータのキャッシュをそれぞれ分離し、チケットシステムの拡張性が向上した。
